R39-DECOMP-VIEWER ②: the clash/QA section out of app.ts — 851 lines
app.ts 5,114 -> 4,272, the pin ratcheted with it. This is the slice that moves
the number: qa was the largest of the four builders in buildToolsPanel.

I HAVE TO CORRECT MYSELF ON THE CENTRAL CLAIM OF ①. I said app.ts has no
parity gate, having checked that no test IMPORTS it. That was true and I
generalised it wrongly: `toolsSplit.test.ts` covers this exact section by
READING ITS SOURCE, and it caught this move immediately — from app.ts alone, 42
inventoried controls had vanished.

So the move did have a real parity gate, and it is a better one than tsc: it
asserts every one of the 42 pre-split controls still exists, that each is on
the side of the seam the split intended, and that the two sections share none.
It went red for the right reason. The property did not change; the file list
did. SECTION_SOURCES now names the extracted modules, so the inventory is
still checked in both directions.

That is the same over-generalisation shape I have been naming all day — a
measurement ("nothing imports it") carried past what it supported ("there is no
gate"). It was caught by running the suite rather than by reasoning, which is
the argument for running it even when you have argued it cannot help.

Evidence the move changed nothing, in order of strength:

  1. The 851 moved lines are BYTE-IDENTICAL to the original. Verified by
     diffing the extracted body against `git show HEAD:app.ts` at the original
     line range — not by reading it. The body is deliberately NOT re-indented:
     re-indenting 851 lines risks silently editing the content of a multi-line
     template literal inside what is meant to be a no-behaviour-change move.
  2. toolsSplit.test.ts passes with the file list updated — all 42 controls
     present, each on the intended side.
  3. tsc clean, which is what forced every capture to be threaded.

tsc earned its keep as promised: it enumerated 60 errors on the first pass and
found FOUR captures my census had missed — container, reloadModelPins,
selectByGuid, waitForPublish, refreshFederation — plus three return types I had
guessed wrong (loadProjectModel returns Promise<boolean>, not void). Every one
would have been a runtime failure.

selectedGuid is passed as an ACCESSOR, not a value. It is a `let` in app.ts, so
a value would freeze the selection at panel-build time and every handler in
this 851-line section would act on a stale GlobalId, silently, forever. That is
the one class tsc cannot see, removed by construction.

Two mutations checked red: dropping an inventoried control from the extracted
file, and forgetting to list the extracted file in SECTION_SOURCES (which would
silently shrink the population the gate reads — this file's own failure mode,
one level up).

1307/1307 web green (116 files), tsc + eslint clean, test_file_sizes OK.
